requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
discord.py>=2.3.0

# Platform-specific (install as needed)
# python-telegram-bot  # For Telegram
# yowsup2  # For WhatsApp